Generate time series/longitudinal COVID datasets (#1685)
Generates _time_series versions of each COVID dataset. These datasets have all the same fields as the originals but are also broken down on a time_period variable.

The time_period variable is the year and month fields of the cdc_case_earliest_dt variable for the case from the original data. So this means that each case, death and hospitalization is being counted as part of the year and month the case was first reported. It is imported to note then that the deaths ARE NOT reported in the month the death occurred, but rather the month that that case was first reported.

Address the backend component of: #1703